artifacts/api-server-python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-multipart==0.0.6
python-dateutil==2.8.2
gunicorn==21.2.0
openai>=1.0.0
python-dotenv>=1.0.0
